+++ script/conf.sh
@@ -1,7 +1,7 @@
 Target="v2ray-web"
 Docker="king011/v2ray-web"
 Dir=$(cd "$(dirname $BASH_SOURCE)/.." && pwd)
-Version="v1.7.3"
+Version="v1.7.4"
 View=1
 Platforms=(
     darwin/amd64